+++ mappers/MMC3.h
@@ -12,23 +12,23 @@ public:
 	std::string name() const override;
 
 public:
-	uint8_t read_6(uint32_t address) override;
-	uint8_t read_7(uint32_t address) override;
+	uint8_t read_6(uint_least16_t address) override;
+	uint8_t read_7(uint_least16_t address) override;
 
 public:
-	void write_6(uint32_t address, uint8_t value) override;
-	void write_7(uint32_t address, uint8_t value) override;
-	void write_8(uint32_t address, uint8_t value) override;
-	void write_9(uint32_t address, uint8_t value) override;
-	void write_a(uint32_t address, uint8_t value) override;
-	void write_b(uint32_t address, uint8_t value) override;
-	void write_c(uint32_t address, uint8_t value) override;
-	void write_d(uint32_t address, uint8_t value) override;
-	void write_e(uint32_t address, uint8_t value) override;
-	void write_f(uint32_t address, uint8_t value) override;
+	void write_6(uint_least16_t address, uint8_t value) override;
+	void write_7(uint_least16_t address, uint8_t value) override;
+	void write_8(uint_least16_t address, uint8_t value) override;
+	void write_9(uint_least16_t address, uint8_t value) override;
+	void write_a(uint_least16_t address, uint8_t value) override;
+	void write_b(uint_least16_t address, uint8_t value) override;
+	void write_c(uint_least16_t address, uint8_t value) override;
+	void write_d(uint_least16_t address, uint8_t value) override;
+	void write_e(uint_least16_t address, uint8_t value) override;
+	void write_f(uint_least16_t address, uint8_t value) override;
 
 public:
-	void vram_change_hook(uint32_t vram_address) override;
+	void vram_change_hook(uint_least16_t vram_address) override;
 
 private:
 	int prg_bank(int bank) const;
@@ -46,7 +46,7 @@ protected:
 	uint8_t prg_bank_[2] = {};
 
 	uint64_t prev_ppu_cycle_    = 0;
-	uint32_t prev_vram_address_ = 0xffff;
+	uint_least16_t prev_vram_address_ = 0xffff;
 	uint8_t command_            = 0;
 	uint8_t irq_latch_          = 0;
 	uint8_t irq_counter_        = 0;
