+++ mappers/Mapper001.h
@@ -11,22 +11,22 @@ public:
 public:
 	std::string name() const override;
 
-	uint8_t read_6(uint32_t address) override;
-	uint8_t read_7(uint32_t address) override;
+	uint8_t read_6(uint_least16_t address) override;
+	uint8_t read_7(uint_least16_t address) override;
 
-	void write_6(uint32_t address, uint8_t value) override;
-	void write_7(uint32_t address, uint8_t value) override;
-	void write_8(uint32_t address, uint8_t value) override;
-	void write_9(uint32_t address, uint8_t value) override;
-	void write_a(uint32_t address, uint8_t value) override;
-	void write_b(uint32_t address, uint8_t value) override;
-	void write_c(uint32_t address, uint8_t value) override;
-	void write_d(uint32_t address, uint8_t value) override;
-	void write_e(uint32_t address, uint8_t value) override;
-	void write_f(uint32_t address, uint8_t value) override;
+	void write_6(uint_least16_t address, uint8_t value) override;
+	void write_7(uint_least16_t address, uint8_t value) override;
+	void write_8(uint_least16_t address, uint8_t value) override;
+	void write_9(uint_least16_t address, uint8_t value) override;
+	void write_a(uint_least16_t address, uint8_t value) override;
+	void write_b(uint_least16_t address, uint8_t value) override;
+	void write_c(uint_least16_t address, uint8_t value) override;
+	void write_d(uint_least16_t address, uint8_t value) override;
+	void write_e(uint_least16_t address, uint8_t value) override;
+	void write_f(uint_least16_t address, uint8_t value) override;
 
 private:
-	void write_handler(uint32_t address, uint8_t value);
+	void write_handler(uint_least16_t address, uint8_t value);
 
 private:
 	uint8_t chr_ram_[0x2000] = {};
