+++ mappers/Mapper011.cpp
@@ -22,63 +22,63 @@ std::string Mapper11::name() const {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper11::write_8(uint32_t address, uint8_t value) {
+void Mapper11::write_8(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper11::write_9(uint32_t address, uint8_t value) {
+void Mapper11::write_9(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper11::write_a(uint32_t address, uint8_t value) {
+void Mapper11::write_a(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper11::write_b(uint32_t address, uint8_t value) {
+void Mapper11::write_b(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper11::write_c(uint32_t address, uint8_t value) {
+void Mapper11::write_c(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper11::write_d(uint32_t address, uint8_t value) {
+void Mapper11::write_d(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper11::write_e(uint32_t address, uint8_t value) {
+void Mapper11::write_e(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper11::write_f(uint32_t address, uint8_t value) {
+void Mapper11::write_f(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper11::write_handler(uint32_t address, uint8_t value) {
+void Mapper11::write_handler(uint_least16_t address, uint8_t value) {
 	(void)address;
 	set_prg_89abcdef(value & 0x3);
 	set_chr_0000_1fff((value >> 4) & 0x0f);
