+++ mappers/Mapper013.cpp
@@ -24,63 +24,63 @@ std::string Mapper13::name() const {
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper13::write_8(uint32_t address, uint8_t value) {
+void Mapper13::write_8(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper13::write_9(uint32_t address, uint8_t value) {
+void Mapper13::write_9(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper13::write_a(uint32_t address, uint8_t value) {
+void Mapper13::write_a(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper13::write_b(uint32_t address, uint8_t value) {
+void Mapper13::write_b(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper13::write_c(uint32_t address, uint8_t value) {
+void Mapper13::write_c(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper13::write_d(uint32_t address, uint8_t value) {
+void Mapper13::write_d(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper13::write_e(uint32_t address, uint8_t value) {
+void Mapper13::write_e(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper13::write_f(uint32_t address, uint8_t value) {
+void Mapper13::write_f(uint_least16_t address, uint8_t value) {
 	write_handler(address, value);
 }
 
 //------------------------------------------------------------------------------
 // Name:
 //------------------------------------------------------------------------------
-void Mapper13::write_handler(uint32_t address, uint8_t value) {
+void Mapper13::write_handler(uint_least16_t address, uint8_t value) {
 	(void)address;
 	set_chr_1000_1fff_ram(chr_ram_, value & 3);
 }
